backend/requirements.txt
# backend/requirements.txt
django~=3.2.0
djangorestframework
mysqlclient
django-cors-headers

# Authentication & Security
djangorestframework-simplejwt 
python-dotenv               